refactor: 이미지 생성 프롬프트 ImagePromptBuilder로 분리 (#78)
- GeminiImageGenerator 인라인 prompt 문자열을 ImagePromptBuilder.build()로 이동
- GeminiImageGenerator가 생성자로 주입받아 사용
